// src/app/models/types/enumerates.ts
export enum Estates {
  MESSAGE = 'MESSAGE',
  TYPING = 'TYPING',
  JOINED = 'JOINED',
  LEFT = 'LEFT',
  ICECANDIDATE = 'ICE-CANDIDATE',
  HANGUP = 'HANGUP',
  ANSWER = 'ANSWER',
  OFFER = 'OFFER',
}

export enum ErrorState{
  PERMISSIONDENIEDERROR = "PermissionDeniedError",
  SECURITYERROR = "SecurityError",
  NOTFOUNDERROR = "NotFoundError"
}